Extract spokenDuration to a dedicated file
The helper is used in audio, video, media badges, and quoted-message
builders. Moving it out of PlaybackTimer.kt into SpokenDuration.kt
(same package) makes it easier to locate without changing any call
sites.
